Test Config
Fiddled around with trying to make sense of all of the tests for VCPKG before the other installs happen.

Loosing my mind with all of the options and trying to keep them straight while naming files.

Things are a mess and I will pick this up later cause my brain is on fire.